+++ envs/kickstarter/configs/env-local.php
@@ -57,7 +57,7 @@ $config = [
          */
         'cms' => [
             'class' => 'luya\cms\frontend\Module',
-            'enableCompression' => true, // compressing the cms output (removing white spaces and newlines)
+            'contentCompression' => true, // compressing the cms output (removing white spaces and newlines)
         ],
 
         /*